requirements.txt
streamlit~=1.36.0
streamlit-chat
langchain==0.1.15
python-dotenv~=1.0.1
langchain-groq
langchain-google-genai
langchain-groq
python-dotenv~=1.0.1
chromadb
pymupdf>=1.24.0
keybert~=0.8.4
transformers~=4.39.3
google-generativeai~=0.5.0
unstructured[md]
ollama
torch~=2.3.0
pydantic~=2.7.1